add language to component so it updates on re-render
