+++ app/components/UI/SeedPhraseVideo/index.js
@@ -3,7 +3,9 @@ import PropTypes from 'prop-types';
 import { StyleSheet, View } from 'react-native';
 import MediaPlayer from '../../Views/MediaPlayer';
 import scaling from '../../../util/scaling';
-import { video_source_uri, subtitle_source_tracks } from '../../../util/video';
+import { video_source_uri, getSubtitleUri } from '../../../util/video';
+import I18n from '../../../../locales/i18n';
+import { TextTrackType } from 'react-native-video';
 
 const HEIGHT = scaling.scale(240);
 
@@ -17,17 +19,29 @@ const styles = StyleSheet.create({
   },
 });
 
-const SeedPhraseVideo = ({ style, onClose }) => (
-  <View style={styles.videoContainer}>
-    <MediaPlayer
-      onClose={onClose}
-      uri={video_source_uri}
-      style={[styles.mediaPlayer, style]}
-      textTracks={subtitle_source_tracks}
-      selectedTextTrack={{ type: 'index', value: 0 }}
-    />
-  </View>
-);
+const SeedPhraseVideo = ({ style, onClose }) => {
+  const language = I18n.locale.substr(0, 2);
+  const subtitle_source_tracks = [
+    {
+      index: 0,
+      title: `${String(language).toUpperCase()} CC`,
+      language: `${language}`,
+      type: TextTrackType.VTT,
+      uri: getSubtitleUri(language),
+    },
+  ];
+  return (
+    <View style={styles.videoContainer}>
+      <MediaPlayer
+        onClose={onClose}
+        uri={video_source_uri}
+        style={[styles.mediaPlayer, style]}
+        textTracks={subtitle_source_tracks}
+        selectedTextTrack={{ type: 'index', value: 0 }}
+      />
+    </View>
+  );
+};
 
 SeedPhraseVideo.propTypes = {
   style: PropTypes.object,
